Refactor functions and types to ensure proper seperation and purpose
