+++ utils/stringUtils.ts
@@ -1,0 +1,6 @@
+export const capitalizeWords = (str: string): string => {
+  return str
+    .split(" ")
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(" ");
+}
